refactor: modify PostList, PostItem directory
