+++ components/PostList/index.tsx
@@ -1,20 +1,20 @@
 import type { NextPage } from "next";
 
-import { PostData } from "../../../lib/posts";
+import { PostData } from "../../lib/posts";
 
 import PostItem from "./PostItem";
 
 import styles from "./PostList.module.css";
 
 interface PostListProps {
-  allPostsData: PostData[];
+  postList: PostData[];
 }
 
-const PostList: NextPage<PostListProps> = ({ allPostsData }) => {
+const PostList: NextPage<PostListProps> = ({ postList }) => {
   return (
     <div>
       <div className={styles.post_list_wrapper}>
-        {allPostsData.map((postData) => (
+        {postList.map((postData) => (
           <PostItem key={postData.id} postData={postData} />
         ))}
       </div>
